+++ code/c1_writing_suite_1/1_return/data.json
@@ -1,7 +1,7 @@
 {
   "name": "TestReturnCode_Simple",
   "category": ["code", "iterate"],
-  "task": "Return the multiplied number in the function multiply_int in code.py. You can make sure you have correctly done this by running test.py",
+  "task": "Return the multiplied number in the function multiply_int in test_code.py. You can make sure you have correctly done this by running test.py",
   "dependencies": ["TestReadFile"],
   "cutoff": 120,
   "ground": {
