+++ code/c1_writing_suite_1/3_modify/data.json
@@ -1,7 +1,7 @@
 {
   "name": "TestReturnCode_Modify",
   "category": ["code", "iterate"],
-  "task": "Modify the multiply_int function in code.py to be able to pass in a 'multiplier' argument to multiply the 'num' by 'multiplier'. Both arguments are integers. You can make sure you have correctly done this by running test.py",
+  "task": "Modify the multiply_int function in test_code.py to be able to pass in a 'multiplier' argument to multiply the 'num' by 'multiplier'. Both arguments are integers. You can make sure you have correctly done this by running test.py",
   "dependencies": ["TestReturnCode_Write"],
   "cutoff": 120,
   "ground": {
