+++ code/c1_writing_suite_1/4_tests/artifacts_in/testfile.py
@@ -1,5 +1,5 @@
 # mypy: ignore-errors
-from code import multiply_int
+from test_code import multiply_int
 
 
 def test_multiply_int(num: int, multiplier, expected_result: int) -> None:
