+++ code/c3_writing_suite_2/d3.1_three_sum/data.json
@@ -1,7 +1,7 @@
 {
   "name": "TestThreeSum",
   "category": ["code", "iterate"],
-  "task": "Create a three_sum function in a file called code.py. Given an array of integers, return indices of the three numbers such that they add up to a specific target. You may assume that each input would have exactly one solution, and you may not use the same element twice. Example: Given nums = [2, 7, 11, 15], target = 20, Because nums[0] + nums[1] + nums[2] = 2 + 7 + 11 = 20, return [0, 1, 2].",
+  "task": "Create a three_sum function in a file called test_code.py. Given an array of integers, return indices of the three numbers such that they add up to a specific target. You may assume that each input would have exactly one solution, and you may not use the same element twice. Example: Given nums = [2, 7, 11, 15], target = 20, Because nums[0] + nums[1] + nums[2] = 2 + 7 + 11 = 20, return [0, 1, 2].",
   "dependencies": ["TestFunctionCodeGeneration"],
   "cutoff": 60,
   "ground": {
