+++ code/c3_writing_suite_2/d3_two_sum/data.json
@@ -1,7 +1,7 @@
 {
   "name": "TestFunctionCodeGeneration",
   "category": ["code"],
-  "task": "Create a two_sum function in a file called code.py. Given an array of integers, return indices of the two numbers such that they add up to a specific target. You may assume that each input would have exactly one solution, and you may not use the same element twice. Example: Given nums = [2, 7, 11, 15], target = 9, Because nums[0] + nums[1] = 2 + 7 = 9, return [0, 1].",
+  "task": "Create a two_sum function in a file called test_code.py. Given an array of integers, return indices of the two numbers such that they add up to a specific target. You may assume that each input would have exactly one solution, and you may not use the same element twice. Example: Given nums = [2, 7, 11, 15], target = 9, Because nums[0] + nums[1] = 2 + 7 = 9, return [0, 1].",
   "dependencies": ["TestReturnCode_Write"],
   "cutoff": 90,
   "ground": {
