+++ code/c4_writing_cli_suite_3/2_file_organizer/data.json
@@ -15,7 +15,7 @@
   },
   "info": {
     "difficulty": "basic",
-    "description": "Tests ability for the agent to create a random password generator.",
+    "description": "Tests ability for the agent to code a file organizer.",
     "side_effects": []
   }
 }
